Improve accuracy of ping measurement and calculations

+++ emulinker/src/main/java/org/emulinker/kaillera/controller/v086/action/LoginAction.kt
@@ -45,7 +45,7 @@ class LoginAction : V086Action<UserInformation>, V086ServerEventHandler<UserJoin
           clientHandler.nextMessageNumber,
           user.name!!,
           user.id,
-          user.ping.toLong(),
+          user.ping,
           user.connectionType
         )
       )
